Improve C->K conversion accuracy
Handle conversions of floating point temperatures. The current code
drops the fractional part of a Celsius temperature before making the
conversion to Kelvin. This is a problem when converting from F->C->K,
because the F->C conversion is likely to end up with a floating point
value. When the fractional part is dropped, multiple F temperatures will
end up mapping to the same K temperature.