Make the API work in for Chinese accounts
For Chinese accounts, Dyson has a different API endpoint at
appapi.cp.dyson.cn, and the one ending with "com" won't work.
This patch add an extra check for CN accounts and redirect the call to
the other API endpoint. It's also easier to add more exceptions (if any)
in the future.